# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Healthonyx API
  version: 0.1.0
  description: |
    Contract-first API surface for the current Healthonyx backend.
    This file mirrors the active route wiring in backend/main.go and is intended
    to be updated as part of each backend endpoint change.
servers:
  - url: http://localhost:8080
    description: Local backend
...
paths:
...
  /api/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/me:
    get:
      tags: [Auth]
      summary: Current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/bootstrap:
    get:
      tags: [Bootstrap]
      summary: Bootstrap application data for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bootstrap payload
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List notifications for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification list
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/notifications/preferences:
    get:
      tags: [Notifications]
      summary: Get notification preferences for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification preferences
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Notifications]
      summary: Update notification preferences for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Preferences updated
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/provider-callbacks/sendgrid:
    post:
      tags: [Notifications]
      summary: Ingest SendGrid delivery callbacks
      responses:
        "200":
          description: Callback processed
        "401":
          description: Invalid webhook signature
  /api/provider-callbacks/twilio:
    post:
      tags: [Notifications]
      summary: Ingest Twilio delivery callbacks
      responses:
        "200":
          description: Callback processed
        "401":
          description: Invalid webhook signature
  /api/patient/dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Patient dashboard summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/doctor/dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Doctor dashboard summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/doctor/critical-escalations:
    get:
      tags: [Dashboard]
      summary: List and refresh critical-result escalations for current doctor
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Escalation list
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/doctor/critical-escalations/{id}/ack:
    post:
      tags: [Dashboard]
      summary: Acknowledge a critical-result escalation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Escalation acknowledged
        "404":
          $ref: "#/components/responses/NotFound"
  /api/doctor/documents:
    get:
      tags: [Documents]
      summary: List doctor-visible patient documents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Documents list
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/doctor/documents/{id}:
    get:
      tags: [Documents]
      summary: Get doctor-visible patient document details
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Document detail
        "404":
          $ref: "#/components/responses/NotFound"
  /api/doctor/documents/{id}/summarize:
    post:
      tags: [Documents]
      summary: Trigger doctor document summary generation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "202":
          description: Summary request accepted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin:
    get:
      tags: [Admin]
      summary: Admin validation endpoint
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin access granted
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/audit-log:
    get:
      tags: [Admin]
      summary: List audit log entries
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Audit log list
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/user-lifecycle/settings-kpis:
    get:
      tags: [Admin]
      summary: Get admin user lifecycle settings and KPI snapshot
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lifecycle settings and KPIs
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/user-lifecycle/settings:
    put:
      tags: [Admin]
      summary: Update admin user lifecycle settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lifecycle settings updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/user-lifecycle/users:
    get:
      tags: [Admin]
      summary: List users for admin lifecycle management
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin]
      summary: Create user from admin lifecycle console
      security:
        - bearerAuth: []
      responses:
        "201":
          description: User created
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/user-lifecycle/users/{id}:
    put:
      tags: [Admin]
      summary: Update user email/role from admin lifecycle console
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: User updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/user-lifecycle/users/{id}/deactivate:
    patch:
      tags: [Admin]
      summary: Deactivate user account
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: User deactivated
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/user-lifecycle/users/{id}/reset-password:
    post:
      tags: [Admin]
      summary: Reset user password
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Password reset completed
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/ai/observability:
    get:
      tags: [Admin]
      summary: Get AI runtime settings, observability, runtime health, and SLO telemetry
      security:
        - bearerAuth: []
      responses:
        "200":
          description: AI observability snapshot
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/ai/settings:
    put:
      tags: [Admin]
      summary: Update AI runtime settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: AI runtime settings updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/ai/eval:
    post:
      tags: [Admin]
      summary: Run AI evaluation baseline
      security:
        - bearerAuth: []
      responses:
        "200":
          description: AI eval result
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/fhir/export:
    get:
      tags: [Admin]
      summary: Export canonical Healthonyx data as FHIR R4 bundle
      security:
        - bearerAuth: []
      responses:
        "200":
          description: FHIR export bundle
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/fhir/import:
    post:
      tags: [Admin]
      summary: Validate and import FHIR R4 bundle into canonical adapters
      security:
        - bearerAuth: []
      responses:
        "200":
          description: FHIR import validation/apply result
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/notifications:
    get:
      tags: [Admin]
      summary: List admin notifications log
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin notification list
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/notifications/summary:
    get:
      tags: [Admin]
      summary: Get admin notifications summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin notification summary
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/notifications/consent-history:
    get:
      tags: [Admin]
      summary: List notification consent provenance events
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification consent history list
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/notifications/{id}/retry:
    post:
      tags: [Admin]
      summary: Retry failed notification by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Retry accepted
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/knowledge-docs:
    get:
      tags: [Admin]
      summary: List knowledge docs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Knowledge docs list
    post:
      tags: [Admin]
      summary: Create knowledge doc
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Knowledge doc created
  /api/admin/knowledge-docs/{id}:
    get:
      tags: [Admin]
      summary: Get knowledge doc
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Knowledge doc details
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Admin]
      summary: Patch knowledge doc title/body/review interval
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Knowledge doc updated
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/knowledge-docs/similarity-scan:
    get:
      tags: [Admin]
      summary: Scan knowledge docs for similarity/conflict candidates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Similarity scan result
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/knowledge-docs/{id}/embeddings:
    post:
      tags: [Admin]
      summary: Upsert embeddings for a knowledge document
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Embeddings updated
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/knowledge-docs/{id}/versions:
    get:
      tags: [Admin]
      summary: List knowledge document versions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Version history
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/knowledge-docs/{id}/review:
    post:
      tags: [Admin]
      summary: Mark knowledge document as reviewed
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Review timestamp updated
        "404":
          $ref: "#/components/responses/NotFound"
  /api/doctor:
    get:
      tags: [Auth]
      summary: Doctor role validation endpoint
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Doctor access granted
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/patient:
    get:
      tags: [Auth]
      summary: Patient role validation endpoint
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Patient access granted
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/geocode:
    get:
      tags: [GeoBooking]
      summary: Geocode patient search query
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Geocode result list
  /api/hospitals/near:
    get:
      tags: [GeoBooking]
      summary: List nearby hospitals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Nearby hospitals
  /api/doctors/search:
    get:
      tags: [GeoBooking]
      summary: Search doctors
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Doctor list
  /api/doctors/{id}/slots:
    get:
      tags: [GeoBooking]
      summary: List open slots for doctor
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Slot list
  /api/doctor/slots:
    post:
      tags: [GeoBooking]
      summary: Create doctor slot
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Slot created
  /api/doctor/slots/{id}:
    delete:
      tags: [GeoBooking]
      summary: Delete doctor slot
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Slot deleted
  /api/appointments/book-slot:
    post:
      tags: [Appointments]
      summary: Book appointment from open slot
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Appointment booked
  /api/appointments:
    post:
      tags: [Appointments]
      summary: Create appointment
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Appointment created
  /api/patient/appointments/{id}/cancel:
    post:
      tags: [Appointments]
      summary: Patient cancel appointment
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Appointment canceled
  /api/appointments/patient:
    get:
      tags: [Appointments]
      summary: List patient appointments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Appointment list
  /api/appointments/doctor:
    get:
      tags: [Appointments]
      summary: List doctor appointments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Appointment list
  /api/appointments/{id}:
    get:
      tags: [Appointments]
      summary: Get appointment by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Appointment detail
        "404":
          $ref: "#/components/responses/NotFound"
  /api/appointments/{id}/status:
    patch:
      tags: [Appointments]
      summary: Update appointment status
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Appointment updated
  /api/appointments/{id}/activity:
    get:
      tags: [Appointments]
      summary: List appointment activity
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Activity list
  /api/appointments/{id}/comments:
    get:
      tags: [Appointments]
      summary: List appointment comments
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Comments list
    post:
      tags: [Appointments]
      summary: Create appointment comment
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "201":
          description: Comment created
  /api/comments/{contextType}/{contextId}:
    get:
      tags: [Appointments]
      summary: List contextual comments by context type and context id
      security:
        - bearerAuth: []
      parameters:
        - name: contextType
          in: path
          required: true
          schema:
            type: string
            enum: [record, document, knowledge_doc]
        - name: contextId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Context comments list
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Appointments]
      summary: Create contextual comment by context type and context id
      security:
        - bearerAuth: []
      parameters:
        - name: contextType
          in: path
          required: true
          schema:
            type: string
            enum: [record, document, knowledge_doc]
        - name: contextId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Context comment created
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/doctors:
    get:
      tags: [GeoBooking]
      summary: List doctors for appointment booking
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Doctor list
  /api/documents:
    get:
      tags: [Documents]
      summary: List patient documents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Document list
    post:
      tags: [Documents]
      summary: Upload patient document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Document uploaded
  /api/documents/{id}/download:
    get:
      tags: [Documents]
      summary: Download patient document
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Document stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /api/patients/{patientId}/files:
    get:
      tags: [PatientFiles]
      summary: List patient files
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PatientIdPath"
      responses:
        "200":
          description: Patient file list
    post:
      tags: [PatientFiles]
      summary: Upload patient file
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PatientIdPath"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                description:
                  type: string
      responses:
        "201":
          description: File uploaded
  /api/files/{id}:
    get:
      tags: [PatientFiles]
      summary: Download patient file
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: File stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
  /api/messages/unread:
    get:
      tags: [Messaging]
      summary: Unread message total for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [patient, doctor, admin]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            role:
              type: string
              enum: [patient, doctor, admin]
...